Fixed compiled error discovered on OS X 64bit (uint and size_t are not compatible) while working on build system. See ticket #1727.
git-svn-id: https://svn.jmodelica.org@3216 7ae51c65-13d5-4921-9d80-6511ddd1419c

+++ trunk/JMI/FMI/XML/src/fmi_xml_variable_list.c
@@ -41,7 +41,7 @@ void fmi_xml_free_variable_list(fmi_xml_variable_list_t* vl) {
 }
 
 /* Get number of variables in a list */
-unsigned int  fmi_xml_get_variable_list_size(fmi_xml_variable_list_t* vl) {
+size_t  fmi_xml_get_variable_list_size(fmi_xml_variable_list_t* vl) {
     return jm_vector_get_size(jm_voidp)(&vl->variables);
 }
